\name{Ksplot}
\alias{Ksplot}
\title{
Draw K-sample plot.
}
\description{
Draw K's plot for function KsamplePlot using KsamplePlot object.
}
\usage{
Ksplot(UseData, Type = "numeric", Method)
}

\arguments{
  \item{UseData}{
KsamplePlot object
}
  \item{Type}{
Type = "numeric" and Type = "binary" when the dependent variable is continuous and binary.
}
  \item{Method}{
The method specified in KsamplePlot
}
}


