\name{Ksplot}
\alias{Ksplot}
\title{
Draw K-sample plot.
}
\description{
Draw K's plot for function KsamplePlot using KsamplePlot object.
}
\usage{
Ksplot(UseDataTest, UseDataTrain, Type = "numeric", Method)
}

\arguments{
  \item{UseDataTest}{
KsamplePlot object of test data.
}
  \item{UseDataTrain}{
KsamplePlot object of training data.
}
  \item{Type}{
Type = "numeric" or Type = "binary" when the dependent variable is continuous or binary.
}
  \item{Method}{
The method specified in KsamplePlot.
}
}


